Move framework specific debug code into relevant case study
